[MOVE/#94] move and add Filtering file
