+++ domain/src/main/java/com/terning/domain/repository/AuthRepository.kt
@@ -4,8 +4,14 @@ import com.terning.domain.entity.request.SignInRequestModel
 import com.terning.domain.entity.response.SignInResponseModel
 
 interface AuthRepository {
+
     suspend fun postSignIn(
         authorization: String,
         request: SignInRequestModel
     ): Result<SignInResponseModel>
+
+    suspend fun postFiltering(
+        userId : Long,
+
+    )
 }
