+++ feature/src/main/java/com/terning/feature/filtering/filtering/FilteringOneScreen.kt
@@ -1,4 +1,4 @@
-package com.terning.feature.onboarding.filtering
+package com.terning.feature.filtering.filtering
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
@@ -7,40 +7,30 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
 import com.terning.core.designsystem.component.button.RectangleButton
-import com.terning.core.designsystem.component.datepicker.DatePickerUI
 import com.terning.core.designsystem.component.image.TerningImage
 import com.terning.core.designsystem.component.topappbar.BackButtonTopAppBar
 import com.terning.core.designsystem.theme.TerningTheme
 import com.terning.feature.R
-import java.util.Calendar
+import com.terning.feature.filtering.filtering.component.StatusOneRadioGroup
+import com.terning.feature.filtering.filtering.navigation.navigateFilteringTwo
 
 @Composable
-fun FilteringThreeScreen(
+fun FilteringOneScreen(
+    name: String,
     navController: NavController,
     modifier: Modifier = Modifier,
     viewModel: FilteringViewModel = hiltViewModel(),
+    onButtonClick: (Int) -> Unit = {},
 ) {
-    val currentYear = Calendar.getInstance().get(Calendar.YEAR)
-    val currentMonth = Calendar.getInstance().get(Calendar.MONTH)
-
-    var chosenYear by remember { mutableIntStateOf(currentYear) }
-    var chosenMonth by remember { mutableIntStateOf(currentMonth) }
-
-    LaunchedEffect(key1 = chosenYear, key2 = chosenMonth) {
-        viewModel.fetchStartYear(chosenYear)
-        viewModel.fetchStartMonth(chosenMonth)
-    }
+    val isButtonValid = remember { mutableStateOf(false) }
 
     Scaffold(
         modifier = modifier,
@@ -56,14 +46,17 @@ fun FilteringThreeScreen(
                 .padding(paddingValues)
         ) {
             TerningImage(
-                painter = R.drawable.ic_filtering_status3,
+                painter = R.drawable.ic_filtering_status1,
                 modifier = modifier.padding(
                     top = 20.dp,
                     start = 24.dp
                 )
             )
             Text(
-                text = stringResource(id = R.string.filtering_status3_title),
+                text = stringResource(
+                    id = R.string.filtering_status1_title,
+                    name
+                ),
                 style = TerningTheme.typography.title3,
                 modifier = modifier.padding(
                     top = 19.dp,
@@ -71,7 +64,10 @@ fun FilteringThreeScreen(
                 )
             )
             Text(
-                text = stringResource(id = R.string.filtering_status3_sub),
+                text = stringResource(
+                    id = R.string.filtering_status1_sub,
+                    name
+                ),
                 style = TerningTheme.typography.body5,
                 modifier = modifier.padding(
                     top = 3.dp,
@@ -79,20 +75,29 @@ fun FilteringThreeScreen(
                     bottom = 25.dp
                 )
             )
-            Spacer(modifier = modifier.weight(1f))
-            DatePickerUI(
-                chosenYear = chosenYear,
-                chosenMonth = chosenMonth,
-                onYearChosen = { chosenYear = it },
-                onMonthChosen = { chosenMonth = it },
+            StatusOneRadioGroup(
+                onButtonClick = { index ->
+                    onButtonClick(index)
+                    isButtonValid.value = true
+                    viewModel.fetchGrade(index)
+                }
             )
-            Spacer(modifier = modifier.weight(3f))
+            Text(
+                text = stringResource(id = R.string.filtering_status1_warning),
+                style = TerningTheme.typography.detail3,
+                modifier = modifier.padding(
+                    start = 24.dp,
+                    top = 9.dp
+                )
+            )
+            Spacer(modifier = modifier.weight(1f))
             RectangleButton(
                 style = TerningTheme.typography.button0,
                 paddingVertical = 20.dp,
                 text = R.string.filtering_button,
-                onButtonClick = { viewModel.postFilteringWithServer() },
+                onButtonClick = { navController.navigateFilteringTwo() },
                 modifier = modifier.padding(bottom = 12.dp),
+                isEnabled = isButtonValid.value
             )
         }
     }
