+++ feature/src/main/java/com/terning/feature/filtering/filtering/FilteringThreeScreen.kt
@@ -1,4 +1,4 @@
-package com.terning.feature.onboarding.filtering
+package com.terning.feature.filtering.filtering
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
@@ -7,30 +7,40 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
 import com.terning.core.designsystem.component.button.RectangleButton
+import com.terning.core.designsystem.component.datepicker.DatePickerUI
 import com.terning.core.designsystem.component.image.TerningImage
 import com.terning.core.designsystem.component.topappbar.BackButtonTopAppBar
 import com.terning.core.designsystem.theme.TerningTheme
 import com.terning.feature.R
-import com.terning.feature.onboarding.filtering.component.StatusTwoRadioGroup
-import com.terning.feature.onboarding.filtering.navigation.navigateFilteringThree
+import java.util.Calendar
 
 @Composable
-fun FilteringTwoScreen(
+fun FilteringThreeScreen(
     navController: NavController,
     modifier: Modifier = Modifier,
     viewModel: FilteringViewModel = hiltViewModel(),
-    onButtonClick: (Int) -> Unit = {},
 ) {
+    val currentYear = Calendar.getInstance().get(Calendar.YEAR)
+    val currentMonth = Calendar.getInstance().get(Calendar.MONTH)
 
-    val isButtonValid = remember { mutableStateOf(false) }
+    var chosenYear by remember { mutableIntStateOf(currentYear) }
+    var chosenMonth by remember { mutableIntStateOf(currentMonth) }
+
+    LaunchedEffect(key1 = chosenYear, key2 = chosenMonth) {
+        viewModel.fetchStartYear(chosenYear)
+        viewModel.fetchStartMonth(chosenMonth)
+    }
 
     Scaffold(
         modifier = modifier,
@@ -46,14 +56,14 @@ fun FilteringTwoScreen(
                 .padding(paddingValues)
         ) {
             TerningImage(
-                painter = R.drawable.ic_filtering_status2,
+                painter = R.drawable.ic_filtering_status3,
                 modifier = modifier.padding(
                     top = 20.dp,
                     start = 24.dp
                 )
             )
             Text(
-                text = stringResource(id = R.string.filtering_status2_title),
+                text = stringResource(id = R.string.filtering_status3_title),
                 style = TerningTheme.typography.title3,
                 modifier = modifier.padding(
                     top = 19.dp,
@@ -61,7 +71,7 @@ fun FilteringTwoScreen(
                 )
             )
             Text(
-                text = stringResource(id = R.string.filtering_status2_sub),
+                text = stringResource(id = R.string.filtering_status3_sub),
                 style = TerningTheme.typography.body5,
                 modifier = modifier.padding(
                     top = 3.dp,
@@ -69,29 +79,20 @@ fun FilteringTwoScreen(
                     bottom = 25.dp
                 )
             )
-            StatusTwoRadioGroup(
-                onButtonClick = {index ->
-                    onButtonClick(index)
-                    isButtonValid.value = true
-                    viewModel.fetchWorkingPeriod(index)
-                }
-            )
-            Text(
-                text = stringResource(id = R.string.filtering_status1_warning),
-                style = TerningTheme.typography.detail3,
-                modifier = modifier.padding(
-                    start = 24.dp,
-                    top = 9.dp
-                )
-            )
             Spacer(modifier = modifier.weight(1f))
+            DatePickerUI(
+                chosenYear = chosenYear,
+                chosenMonth = chosenMonth,
+                onYearChosen = { chosenYear = it },
+                onMonthChosen = { chosenMonth = it },
+            )
+            Spacer(modifier = modifier.weight(3f))
             RectangleButton(
                 style = TerningTheme.typography.button0,
                 paddingVertical = 20.dp,
                 text = R.string.filtering_button,
-                onButtonClick = { navController.navigateFilteringThree() },
+                onButtonClick = { viewModel.postFilteringWithServer() },
                 modifier = modifier.padding(bottom = 12.dp),
-                isEnabled = isButtonValid.value
             )
         }
     }
