+++ feature/src/main/java/com/terning/feature/filtering/filtering/FilteringViewModel.kt
@@ -1,4 +1,4 @@
-package com.terning.feature.onboarding.filtering
+package com.terning.feature.filtering.filtering
 
 import androidx.lifecycle.ViewModel
 import com.terning.domain.repository.AuthRepository
