+++ feature/src/main/java/com/terning/feature/filtering/filtering/navigation/FilteringOneNavigation.kt
@@ -1,4 +1,4 @@
-package com.terning.feature.onboarding.filtering.navigation
+package com.terning.feature.filtering.filtering.navigation
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
@@ -7,7 +7,7 @@ import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import androidx.navigation.toRoute
 import com.terning.core.navigation.Route
-import com.terning.feature.onboarding.filtering.FilteringOneScreen
+import com.terning.feature.filtering.filtering.FilteringOneScreen
 import kotlinx.serialization.Serializable
 
 fun NavController.navigateFilteringOne(
