+++ feature/src/main/java/com/terning/feature/filtering/filtering/navigation/ThreeNavigation.kt
@@ -1,4 +1,4 @@
-package com.terning.feature.onboarding.filtering.navigation
+package com.terning.feature.filtering.filtering.navigation
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
@@ -6,7 +6,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.terning.core.navigation.Route
-import com.terning.feature.onboarding.filtering.FilteringThreeScreen
+import com.terning.feature.filtering.filtering.FilteringThreeScreen
 import kotlinx.serialization.Serializable
 
 fun NavController.navigateFilteringThree(navOptions: NavOptions? = null) {
