+++ feature/src/main/java/com/terning/feature/filtering/startfiltering/navigation/StartFilteringNavigation.kt
@@ -1,4 +1,4 @@
-package com.terning.feature.onboarding.startfiltering.navigation
+package com.terning.feature.filtering.startfiltering.navigation
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
@@ -7,7 +7,7 @@ import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import androidx.navigation.toRoute
 import com.terning.core.navigation.Route
-import com.terning.feature.onboarding.startfiltering.StartFilteringScreen
+import com.terning.feature.filtering.startfiltering.StartFilteringScreen
 import kotlinx.serialization.Serializable
 
 fun NavController.navigateStartFiltering(
