+++ feature/src/main/java/com/terning/feature/filtering/starthome/navigation/StartHomeNavigation.kt
@@ -1,4 +1,4 @@
-package com.terning.feature.onboarding.starthome.navigation
+package com.terning.feature.filtering.starthome.navigation
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
@@ -6,7 +6,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.terning.core.navigation.Route
-import com.terning.feature.onboarding.starthome.StartHomeScreen
+import com.terning.feature.filtering.starthome.StartHomeScreen
 import kotlinx.serialization.Serializable
 
 fun NavController.navigateStartHome(navOptions: NavOptions? = null) {
