+++ feature/src/main/java/com/terning/feature/main/MainNavigator.kt
@@ -12,8 +12,6 @@ import androidx.navigation.navOptions
 import com.terning.feature.calendar.navigation.navigateCalendar
 import com.terning.feature.home.home.navigation.navigateHome
 import com.terning.feature.mypage.navigation.navigateMyPage
-import com.terning.feature.onboarding.filtering.navigation.FilteringOne
-import com.terning.feature.onboarding.signin.navigation.SignIn
 import com.terning.feature.onboarding.signup.navigation.SignUp
 import com.terning.feature.search.search.navigation.navigateSearch
