+++ feature/src/main/java/com/terning/feature/main/MainScreen.kt
@@ -25,13 +25,13 @@ import com.terning.feature.home.changefilter.navigation.changeFilterNavGraph
 import com.terning.feature.home.home.navigation.homeNavGraph
 import com.terning.feature.intern.navigation.internNavGraph
 import com.terning.feature.mypage.navigation.myPageNavGraph
-import com.terning.feature.onboarding.filtering.navigation.filteringOneNavGraph
-import com.terning.feature.onboarding.filtering.navigation.filteringThreeNavGraph
-import com.terning.feature.onboarding.filtering.navigation.filteringTwoNavGraph
+import com.terning.feature.filtering.filtering.navigation.filteringOneNavGraph
+import com.terning.feature.filtering.filtering.navigation.filteringThreeNavGraph
+import com.terning.feature.filtering.filtering.navigation.filteringTwoNavGraph
 import com.terning.feature.onboarding.signin.navigation.signInNavGraph
 import com.terning.feature.onboarding.signup.navigation.signUpNavGraph
-import com.terning.feature.onboarding.startfiltering.navigation.startFilteringNavGraph
-import com.terning.feature.onboarding.starthome.navigation.startHomeNavGraph
+import com.terning.feature.filtering.startfiltering.navigation.startFilteringNavGraph
+import com.terning.feature.filtering.starthome.navigation.startHomeNavGraph
 import com.terning.feature.search.search.navigation.searchNavGraph
 import com.terning.feature.search.searchprocess.navigation.searchProcessNavGraph
